// tests/api/generated/client/types.gen.ts
export type ErrorResponse = {
    message: string;
};

export type TransactionType = {
    id: string;
    name: string;
};

export type Transaction = {
    id: string;
    item: string;
    amount: number;
    date: string;
    subcategoryId: string;
};

export type TransactionRequest = {
    item: string;
    amount: number;
    date: string;
    subcategoryId: string;
};

export type SuccessResponse = {
    message: string;
};

export type Template = {
    id: string;
    name: string;
    itemName: string;
    amount: number | null;
    date: string | null;
    subcategoryId: string;
};

export type TemplateRequest = {
    name: string;
    itemName: string;
    amount: number | null;
    date: string | null;
    subcategoryId: string;
};

export type Subcategory = {
    id: string;
    name: string;
    mainCategoryId: string;
    expenseTypeId: string;
};

export type SubcategoryRequest = {
    name: string;
    mainCategoryId: string;
    expenseTypeId: string;
};

export type MainCategory = {
    id: string;
    name: string;
    expenseTypeId: string;
    transactionTypeId: string;
};

export type MainCategoryRequest = {
    name: string;
    expenseTypeId: string;
    transactionTypeId: string;
};

export type LoginSuccessResponse = {
    role: 'user' | 'caregiver';
    token: string;
    user: {
        email: string;
        name: string;
        id: string;
    };
};

export type LoginRequest = {
    email: string;
    password: string;
};

export type ExpenseType = {
    id: string;
    name: string;
};